backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
httpx==0.25.1
base58==2.1.1
solders==0.18.1
python-multipart==0.0.6